requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-generativeai==0.3.2
httpx==0.27.2
python-dotenv==1.0.1
sse-starlette==2.1.3
sqlalchemy==2.0.36
typing-extensions==4.12.2
google-cloud-aiplatform==1.42.1
google-auth==2.28.1